chore: consolidate parsePositiveInt to single home
