+++ skills/relay-merge/scripts/relay-reconcile-artifact.js
@@ -24,6 +24,7 @@
 const path = require("path");
 const {
   getExpectedManifestRepoRoot,
+  parsePositiveInt,
   validateManifestPaths,
 } = require("../../relay-dispatch/scripts/manifest/paths");
 const {
@@ -74,17 +75,6 @@ function requireNonEmptyArg(flag, label) {
   return value.trim();
 }
 
-function parsePositiveInt(value, label) {
-  if (typeof value !== "string" || !value.trim()) {
-    throw new Error(`${label} is required`);
-  }
-  const parsed = Number(value);
-  if (!Number.isInteger(parsed) || parsed <= 0) {
-    throw new Error(`${label} must be a positive integer`);
-  }
-  return parsed;
-}
-
 function sameBootstrapExemption(data, { artifactPath, writerPr, reason }) {
   const existing = data?.bootstrap_exempt || {};
   return (
@@ -97,7 +87,7 @@ function sameBootstrapExemption(data, { artifactPath, writerPr, reason }) {
 
 function main() {
   const artifactPath = requireNonEmptyArg("--artifact-path", "--artifact-path <path>");
-  const writerPr = parsePositiveInt(getArg("--writer-pr"), "--writer-pr <int>");
+  const writerPr = parsePositiveInt(requireNonEmptyArg("--writer-pr", "--writer-pr <int>"), "--writer-pr <int>");
   const reason = requireNonEmptyArg("--reason", "--reason <text>");
   const skipReviewReason = getArg("--skip-review");
   if (hasFlag("--skip-review") && !String(skipReviewReason || "").trim()) {
