+++ skills/relay-review/scripts/review-runner/common.js
@@ -14,15 +14,6 @@ const git = (repoPath, ...gitArgs) => execGit(repoPath, gitArgs);
 
 const RUBRIC_PASS_THROUGH_STATES = new Set(["loaded"]);
 
-function parsePositiveInt(value, label) {
-  if (value === undefined) return undefined;
-  const parsed = Number(value);
-  if (!Number.isInteger(parsed) || parsed <= 0) {
-    throw new Error(`${label} must be a positive integer`);
-  }
-  return parsed;
-}
-
 function readText(filePath) {
   return fs.readFileSync(filePath, "utf-8");
 }
@@ -35,7 +26,6 @@ function writeText(filePath, text) {
 module.exports = {
   gh,
   git,
-  parsePositiveInt,
   readText,
   RUBRIC_PASS_THROUGH_STATES,
   writeText,
