+++ skills/relay-review/scripts/review-runner/context.js
@@ -4,11 +4,12 @@ const path = require("path");
 const {
   getCanonicalRepoRoot,
   getExpectedManifestRepoRoot,
+  parsePositiveInt,
   validateManifestPaths,
 } = require("../../../relay-dispatch/scripts/manifest/paths");
 const { resolveManifestRecord } = require("../../../relay-dispatch/scripts/relay-resolver");
 const { getRubricAnchorStatus } = require("../../../relay-dispatch/scripts/manifest/rubric");
-const { gh, parsePositiveInt, readText, RUBRIC_PASS_THROUGH_STATES } = require("./common");
+const { gh, readText, RUBRIC_PASS_THROUGH_STATES } = require("./common");
 
 // DNS hostname validation — conservative label allowlist. Rejects leading
 // dashes (which could be interpreted as flags by some CLI tools), whitespace,
